refactor(server): extract handle_init / load_seed_from_data to ws_session.py
The two largest closures inside `websocket_endpoint` (~150 lines combined)
move to standalone async functions in ws_session.py. Both take an explicit
`Connection` plus the loaded engine refs as params — no closure capture.

handle_init's recorder-lifecycle inline now delegates to
`conn.start_action_log_segment(world_engine)` etc. (lifted last commit),
keeping the body tight.

Updated call sites:
- pre-init handshake: `await handle_init(conn, world_engine, safety_checker, session, req)`
- game-loop receiver: same with `is_game_loop=True`

Net ~150 lines removed from the endpoint; the remaining big closures are
the receiver and generator. server.py drops to ~1340 lines.

Refs #78.
