+++ server-components/ws_session.py
@@ -19,25 +19,36 @@ rules in pyproject.toml fire on this code. Keep it that way.
 """
 
 import asyncio
+import base64
+import hashlib
+import io
 import logging
 import threading
+import time
 from dataclasses import dataclass, field
 from queue import Queue
 from typing import TYPE_CHECKING
 
 from fastapi import WebSocket
+from PIL import Image
 from pydantic import BaseModel
 
 from action_logger import ActionLogger
 from app_state import AppState
-from progress_stages import Stage
-from protocol import MessageId, StatusMessage, WarningMessage
+from progress_stages import SESSION_WAITING_FOR_SEED, Stage
+from protocol import InitRequest, MessageId, StatusMessage, WarningMessage
+from safety_cache import save_safety_cache
 from video_recorder import RecordingProperties, VideoRecorder
 
 if TYPE_CHECKING:
-    from engine_manager import WorldEngineManager
+    from engine_manager import Session, WorldEngineManager
+    from safety import SafetyChecker
 
 logger = logging.getLogger(__name__)
+
+
+def _compute_bytes_hash(data: bytes) -> str:
+    return hashlib.sha256(data).hexdigest()
 
 
 @dataclass
@@ -206,6 +217,175 @@ class Connection:
             self.main_loop.call_soon_threadsafe(self.frame_ready.set)
         except Exception:
             pass
+
+
+async def load_seed_from_data(
+    conn: Connection,
+    world_engine: "WorldEngineManager",
+    safety_checker: "SafetyChecker",
+    image_data_b64: str | None,
+    seed_filename: str | None = None,
+) -> bool:
+    """Validate safety and load a seed from base64 image data.
+
+    Returns True iff the seed was loaded (or already loaded and matched).
+    Failure paths surface a typed warning over the websocket and return False.
+    """
+    if not image_data_b64:
+        logger.warning(f"[{conn.client_host}] Missing seed image data")
+        await conn.send_warning(MessageId.SEED_MISSING_DATA)
+        return False
+
+    try:
+        image_bytes = base64.b64decode(image_data_b64)
+    except Exception as e:
+        logger.warning(f"[{conn.client_host}] Invalid base64 seed data: {e}")
+        await conn.send_warning(MessageId.SEED_INVALID_DATA)
+        return False
+
+    img_hash = _compute_bytes_hash(image_bytes)
+
+    # Same seed already loaded? Skip the safety inference + load entirely.
+    if img_hash == conn.current_seed_hash:
+        logger.info(f"[{conn.client_host}] Seed unchanged (hash match), skipping reload")
+        return True
+
+    # Safety check (cache hit, then live inference if missing)
+    cache = conn.state.safety_hash_cache
+    if img_hash in cache:
+        cached = cache[img_hash]
+        if not cached.get("is_safe", False):
+            logger.warning(f"[{conn.client_host}] Seed marked as unsafe (cached)")
+            await conn.send_warning(MessageId.SEED_UNSAFE)
+            return False
+    else:
+        try:
+
+            def _check_safety():
+                pil_img = Image.open(io.BytesIO(image_bytes)).convert("RGB")
+                return safety_checker.check_pil_image(pil_img)
+
+            safety_result = await asyncio.to_thread(_check_safety)
+        except Exception as e:
+            logger.warning(f"[{conn.client_host}] Safety check failed: {e}")
+            await conn.send_warning(MessageId.SEED_SAFETY_CHECK_FAILED)
+            return False
+
+        is_safe = safety_result.get("is_safe", False)
+        cache[img_hash] = {
+            "is_safe": is_safe,
+            "scores": safety_result.get("scores", {}),
+            "checked_at": time.time(),
+        }
+        save_safety_cache(cache)
+
+        if not is_safe:
+            logger.warning(f"[{conn.client_host}] Seed marked as unsafe")
+            await conn.send_warning(MessageId.SEED_UNSAFE)
+            return False
+
+    # Load the seed onto the engine
+    display_name = seed_filename or img_hash[:12]
+    logger.info(f"[{conn.client_host}] Loading seed '{display_name}'")
+    loaded_frame = await world_engine.load_seed_from_base64(image_data_b64)
+    if loaded_frame is None:
+        logger.error(f"[{conn.client_host}] Failed to load seed")
+        await conn.send_warning(MessageId.SEED_LOAD_FAILED)
+        return False
+
+    world_engine.seed_frame = loaded_frame
+    world_engine.original_seed_frame = loaded_frame
+    conn.current_seed_hash = img_hash
+    conn.current_seed_filename = seed_filename
+    logger.info(f"[{conn.client_host}] Seed loaded successfully")
+    return True
+
+
+async def handle_init(
+    conn: Connection,
+    world_engine: "WorldEngineManager",
+    safety_checker: "SafetyChecker",
+    session: "Session",
+    req: InitRequest,
+    *,
+    is_game_loop: bool = False,
+) -> tuple[bool, bool]:
+    """Apply an InitRequest's deltas to the connection / engine.
+
+    Returns `(ready, seed_loaded)`: `ready` means the session has a
+    seed frame and can begin generating; `seed_loaded` means a fresh
+    seed was applied in this call.
+
+    `model_fields_set` distinguishes "field absent" from "field present
+    and explicitly None", so partial updates (`{"action_logging": false}`
+    while leaving everything else untouched) work as the renderer expects.
+    """
+    present = req.model_fields_set
+    model_uri = (req.model or "").strip()
+    seed_data = req.seed_image_data
+    seed_filename = req.seed_filename
+    quant = req.quant
+
+    # Update flags
+    if "scene_authoring" in present and req.scene_authoring is not None:
+        conn.scene_authoring_requested = req.scene_authoring
+    if "action_logging" in present and req.action_logging is not None:
+        conn.action_logging_requested = req.action_logging
+    if "video_recording" in present and req.video_recording is not None:
+        conn.video_recording_requested = req.video_recording
+    if "video_output_dir" in present:
+        conn.video_output_dir = req.video_output_dir
+    if "biome_version" in present:
+        conn.biome_version = req.biome_version
+    if "cap_inference_fps" in present and req.cap_inference_fps is not None:
+        conn.cap_inference_fps = req.cap_inference_fps
+
+    # Sync recorder lifecycle with requested state during gameplay
+    if is_game_loop:
+        if conn.action_logging_requested and conn.action_logger is None:
+            conn.action_logger = ActionLogger(conn.client_host)
+            conn.start_action_log_segment(world_engine)
+            logger.info(f"[{conn.client_host}] Action logging enabled")
+        elif not conn.action_logging_requested and conn.action_logger is not None:
+            conn.action_logger.end_segment()
+            conn.action_logger = None
+            logger.info(f"[{conn.client_host}] Action logging disabled")
+
+        if conn.video_recording_requested and conn.video_recorder is None:
+            conn.start_video_segment(world_engine)
+            logger.info(f"[{conn.client_host}] Video recording enabled")
+        elif not conn.video_recording_requested and conn.video_recorder is not None:
+            conn.video_recorder.end_segment()
+            conn.video_recorder = None
+            logger.info(f"[{conn.client_host}] Video recording disabled")
+
+    # Model delta — reload if model URI or quantization changed.
+    # The engine must be loaded before the seed so that seed_target_size
+    # and temporal_compression are resolved from the actual model config.
+    model_changed = False
+    quant_changed = "quant" in present and quant != world_engine.quant
+    if model_uri and (model_uri != world_engine.model_uri or quant_changed):
+        verb = "Live model switch" if is_game_loop else "Requested model"
+        logger.info(f"[{conn.client_host}] {verb}: {model_uri} (quant={quant})")
+        world_engine.set_progress_callback(conn.push_progress, conn.main_loop)
+        await world_engine.load_engine(model_uri, quant=quant)
+        world_engine.set_progress_callback(None)
+        world_engine.seed_frame = None
+        session.perceptual_frame_count = 0
+        session.max_perceptual_frames = (world_engine.n_frames - 2) * world_engine.temporal_compression
+        model_changed = True
+        logger.info(f"[{conn.client_host}] Model loaded: {world_engine.model_uri}")
+
+    # Seed delta
+    seed_loaded = False
+    if seed_data:
+        seed_loaded = await load_seed_from_data(conn, world_engine, safety_checker, seed_data, seed_filename)
+
+    if model_changed and not seed_loaded and not world_engine.seed_frame:
+        await conn.send_stage(SESSION_WAITING_FOR_SEED)
+
+    ready = seed_loaded or (world_engine.seed_frame is not None)
+    return ready, seed_loaded
 
 
 async def run_sender(conn: Connection) -> None:
